+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - fixed sandman crash issue introduced in 1.7.0
 - fixed trace log filter case case sensitive
 - fixed performance issues with Delete V2
+- fixed issue with NtQueryDirectoryFile data alignment [#2443](https://github.com/sandboxie-plus/Sandboxie/issues/2443)
 
 
 ## [1.7.0 / 5.62.0] - 2022-12-27